refactor(ui): rename useMDContainer to paneScrollContainer for clarity
