+++ src/shared/ui/Button/MDFabContainer.vue
@@ -7,7 +7,7 @@ import { useOverlayContainer } from '../Overlay';
 import { TeleportContainer } from '@shared/lib/teleportContainer';
 import { autoUpdate, offset, shift, useFloating } from '@floating-ui/vue';
 import { useMainContentAriaHidden } from '../AriaHidden';
-import { usePaneScrollContainer } from '../Layout/useMDContainer';
+import { usePaneScrollContainer } from '../Layout/paneScrollContainer';
 
 const props = defineProps<{
   /** Hides the floating action while scrolling down and restores it on upward scroll or focus. */
