+++ src/shared/ui/Layout/index.ts
@@ -1,4 +1,4 @@
 export { default as MDSplitLayout } from './MDSplitLayout.vue';
 export { default as MDPane } from './MDPane.vue';
 export { usePaneContext } from './paneContext';
-export { usePaneScrollContainer } from './useMDContainer';
+export { usePaneScrollContainer } from './paneScrollContainer';
